Refactor conversation and optimize mentions
